Mapper & Passenger post & Passenger GET & need to do the testing of passenger

+++ PersistenceLayer/src/main/java/com/example/TDDAssessmentBackend/persistence/repository/FlightRepository.java
@@ -3,6 +3,6 @@ package com.example.TDDAssessmentBackend.persistence.repository;
 import com.example.TDDAssessmentBackend.persistence.models.Flight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlightRepository extends JpaRepository<Flight, Long> {
+public interface FlightRepository extends JpaRepository<Flight, Integer> {
 
 }
