+++ PersistenceLayer/src/main/java/com/example/TDDAssessmentBackend/persistence/repository/PassengerRepository.java
@@ -3,6 +3,6 @@ package com.example.TDDAssessmentBackend.persistence.repository;
 import com.example.TDDAssessmentBackend.persistence.models.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PassengerRepository extends JpaRepository<Passenger, Long> {
+public interface PassengerRepository extends JpaRepository<Passenger, Integer> {
 
 }
